* Incorporated fixes submitted by Bryan Donlan <bdonlan@gmail.com>:   - Fixed missing return statement after error in S3_list_bucket   - Fixed integer types in string functions   - Added flags to S3_initialize(), and be sure to call curl_global_init()     and curl_global_cleanup() * Changed s3 utility to use path-style URI as default, but to refuse to   create a bucket which fails virtual-host-style validation unless the -f   (--force) option is passed to s3 * Allow S3_initialize() to be called more than once with no side effects, as   long as S3_deinitialize() is called the same number of times when libs3   is being shut down

+++ inc/string_buffer.h
@@ -49,7 +49,7 @@
 #define string_buffer_append(sb, str, len, all_fit)                     \
     do {                                                                \
         sb##Len += snprintf(&(sb[sb##Len]), sizeof(sb) - sb##Len - 1,   \
-                           "%.*s", len, str);                           \
+                            "%.*s", (int) (len), str);                  \
         if (sb##Len > (sizeof(sb) - 1)) {                               \
             sb##Len = sizeof(sb) - 1;                                   \
             all_fit = 0;                                                \
@@ -83,7 +83,7 @@
     do {                                                                \
         smb##Size += (snprintf(&(smb[smb##Size]),                       \
                                sizeof(smb) - smb##Size,                 \
-                               "%.*s", len, str) + 1);                  \
+                               "%.*s", (int) (len), str) + 1);          \
         if (smb##Size > sizeof(smb)) {                                  \
             smb##Size = sizeof(smb);                                    \
             all_fit = 0;                                                \
